refactor: Header, Navber Page에서 분리
