+++ src/index.tsx
@@ -1,8 +1,31 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
+import StopWatchPage from './pages/StopWatchPage';
+import WorldDigaitalClockPage from './pages/WorldDigaitalClockPage';
+import TimerPage from './pages/TimerPage';
 import App from './App';
-import GlobalStyle from './styles/GlobalStyles';
-import Layout from './components/Layout';
+import { createBrowserRouter, RouterProvider } from 'react-router-dom';
+
+const router = createBrowserRouter([
+  {
+    path: '/',
+    element: <App />,
+    children: [
+      {
+        path: '',
+        element: <WorldDigaitalClockPage />,
+      },
+      {
+        path: '/stop-watch',
+        element: <StopWatchPage />,
+      },
+      {
+        path: '/timer',
+        element: <TimerPage />,
+      },
+    ],
+  },
+]);
 
 const root = ReactDOM.createRoot(
   document.getElementById('root') as HTMLElement
@@ -10,9 +33,6 @@ const root = ReactDOM.createRoot(
 
 root.render(
   <React.StrictMode>
-    <GlobalStyle />
-    <Layout>
-      <App />
-    </Layout>
+    <RouterProvider router={router} />
   </React.StrictMode>
 );
